Add files via upload
I spent a couple of hours yesterday with Google Translate (now know how to do it faster) to translate Christian Buil new paper on "How to calculate the spectral response of a spectrograph". Here is the pdf of that document.  If and when I get time I will replace this with a better replication that does a better job of formatting the multiples columns in the document.